function/requirements.txt
# Core dependencies
functions-framework==3.8.1
google-cloud-storage==1.40.0
google-cloud-bigquery==1.44.0
pandas==1.5.2
PyYAML==6.0

# Image analysis dependencies
google-cloud-aiplatform==2.11.0
tensorflow==2.11.0
tensorflow-addons==0.17.0

# Logging
logging==3.6.0

# Utilities
json==2.1.3
ospath==2.3.5
subprocess==3.9.2

# Optional dependencies (consider removing if not used)
bigframes==0.19.2
vertexai==2.11.0
